Stage state
This is the state when I lost interest in this given approach. Sure this
is indeed possible but it will be a lot of work to maintain and I do not
have the time to finish this in a reasonable time.

Instead I will try another alternative approach that may, or may not
make the snap less portable, but I will be easier and quicker to
maintain.