Use lxd remotes instead of local socket.
Looks like it's a little tricky to be allowed to use the lxd interface
in the store. I will use remotes for now, this grants me the same
access, just a little more setup, and there are a few potential
security problems if the user configures the lxd remote incorrectly.

There is a new config value called "lxd-remotes" with the format:
[{"url": "...", "password": "..."}], it's a list but at the moment
I only support one value.