Remove use of lxd interface.
I removed the lxd interface because this was a new use case for
Canonical it's not clear at the moment if they will allow and snap
in the store that uses the lxd interface. I have already changed
the code in earlier commits to use the remote interface of lxd.